update animation so it stops at the right value, yikez
mv include

+++ src/components/rotate.cpp
@@ -5,5 +5,5 @@
 
 void rotate_model(ModelComponent & mc, const RotationComponent & rc) {
         if(rc.axis.x + rc.axis.y + rc.axis.z)
-          mc.matrix = glm::rotate(mc.matrix, rc.radians, rc.axis);
+          mc.matrix = glm::rotate(mc.matrix, (float)rc.radians, rc.axis);
 }
